+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Fixed
 - Fix a bug where deleting a video when the last video in a category is selected breaks things.
+- [Issue 447](https://github.com/aza547/wow-recorder/issues/447) - Fix a bug where could end up in an error state after sleeping Windows.
 
 ## [3.19.2] - 2023-10-23
 ### Fixed